Move test from test_forms into test_email
All the related tests are there, and the test itself fits better there.
